fix(auth_broker): fx-desktop-v1 merged into fx-ios-v1
