+++ app/tests/spec/models/auth_brokers/fx-ios-v1.js
@@ -5,23 +5,24 @@
 define(function (require, exports, module) {
   'use strict';
 
-  const Account = require('models/account');
   const { assert } = require('chai');
   const FxiOSAuthenticationBroker = require('models/auth_brokers/fx-ios-v1');
-  const FxDesktopV1AuthenticationBroker = require('models/auth_brokers/fx-desktop-v1');
   const NullChannel = require('lib/channels/null');
   const Relier = require('models/reliers/relier');
   const sinon = require('sinon');
+  const User = require('models/user');
   const WindowMock = require('../../../mocks/window');
 
   const IMMEDIATE_UNVERIFIED_LOGIN_UA_STRING = 'Mozilla/5.0 (iPhone; CPU iPhone OS 8_3 like Mac OS X) AppleWebKit/600.1.4 (KHTML, like Gecko) FxiOS/6.1 Mobile/12F69 Safari/600.1.4'; //eslint-disable-line max-len
   const CHOOSE_WHAT_TO_SYNC_UA_STRING = 'Mozilla/5.0 (iPhone; CPU iPhone OS 8_3 like Mac OS X) AppleWebKit/600.1.4 (KHTML, like Gecko) FxiOS/11.0 Mobile/12F69 Safari/600.1.4'; //eslint-disable-line max-len
 
   describe('models/auth_brokers/fx-ios-v1', () => {
+    let account;
     let broker;
     let channel;
     const loginMessageDelayMS = 250;
     let relier;
+    let user;
     let windowMock;
     let sandbox;
 
@@ -32,6 +33,14 @@ define(function (require, exports, module) {
         loginMessageDelayMS: loginMessageDelayMS,
         relier: relier,
         window: windowMock
+      });
+
+      user = new User();
+      account = user.initAccount({
+        email: 'testuser@testuser.com',
+        keyFetchToken: 'key-fetch-token',
+        uid: 'uid',
+        unwrapBKey: 'unwrap-b-key'
       });
       sandbox.stub(broker, '_hasRequiredLoginFields').callsFake(() => true);
     }
@@ -80,15 +89,10 @@ define(function (require, exports, module) {
     });
 
     describe('_notifyRelierOfLogin', () => {
-      let account;
-
       beforeEach(() => {
         sandbox.stub(broker, 'send').callsFake(() => Promise.resolve());
         sandbox.spy(windowMock, 'setTimeout');
         sandbox.spy(windowMock, 'clearTimeout');
-        account = new Account({
-          uid: 'uid'
-        });
       });
 
       function testLoginSent(triggerLoginCB) {
@@ -126,22 +130,136 @@ define(function (require, exports, module) {
       });
 
       describe('afterCompleteSignInWithCode', () => {
-        let account;
-
         beforeEach(() => {
           sandbox.spy(broker, 'afterCompleteSignInWithCode');
           sandbox.spy(broker, '_notifyRelierOfLogin');
-          sandbox.spy(FxDesktopV1AuthenticationBroker.prototype, 'afterCompleteSignInWithCode');
-          account = new Account({
-            uid: 'uid'
-          });
+          sandbox.spy(FxiOSAuthenticationBroker.prototype, 'afterCompleteSignInWithCode');
+
           return broker.afterCompleteSignInWithCode(account);
         });
 
         it('broker calls correct methods', () => {
-          assert.isTrue(FxDesktopV1AuthenticationBroker.prototype.afterCompleteSignInWithCode.called);
+          assert.isTrue(broker.afterCompleteSignInWithCode.called);
           assert.isTrue(broker._notifyRelierOfLogin.called);
         });
+      });
+    });
+
+    describe('createChannel', () => {
+      it('creates a channel', () => {
+        assert.ok(broker.createChannel());
+      });
+    });
+
+    describe('channel errors', () => {
+      it('are propagated outwards', () => {
+        var channel = broker.createChannel();
+
+        var errorSpy = sinon.spy();
+        broker.on('error', errorSpy);
+
+        var error = new Error('malformed message');
+        channel.trigger('error', error);
+        assert.isTrue(errorSpy.calledWith(error));
+      });
+    });
+
+    describe('afterLoaded', () => {
+      it('sends a `loaded` message', () => {
+        sinon.spy(broker, 'send');
+
+        return broker.afterLoaded()
+          .then(() => {
+            assert.isTrue(broker.send.calledWith('loaded'));
+          });
+      });
+    });
+
+    describe('afterSignIn', () => {
+      it('notifies the channel of login, halts by default', () => {
+        sinon.spy(broker, 'send');
+
+        return broker.afterSignIn(account)
+          .then((result) => {
+            assert.isTrue(broker.send.calledWith('login'));
+            assert.isTrue(result.halt);
+          });
+      });
+    });
+
+    describe('afterSignInConfirmationPoll', () => {
+      it('navigates to `signin_confirmed` by default', () => {
+        return broker.afterSignInConfirmationPoll(account)
+          .then(() => {
+            assert.equal(broker.getBehavior('afterSignInConfirmationPoll').type, 'navigate');
+            assert.equal(broker.getBehavior('afterSignInConfirmationPoll').endpoint, 'signin_confirmed');
+          });
+      });
+    });
+
+    describe('beforeSignUpConfirmationPoll', () => {
+      it('notifies the channel of login', () => {
+        sinon.spy(broker, 'send');
+
+        return broker.beforeSignUpConfirmationPoll(account)
+          .then(() => {
+            assert.isTrue(broker.send.calledWith('login'));
+          });
+      });
+    });
+
+    describe('afterSignUpConfirmationPoll', () => {
+      it('navigates to `signup_confirmed` by default', () => {
+        return broker.afterSignUpConfirmationPoll(account)
+          .then(() => {
+            assert.equal(broker.getBehavior('afterSignUpConfirmationPoll').type, 'navigate');
+            assert.equal(broker.getBehavior('afterSignUpConfirmationPoll').endpoint, 'signup_confirmed');
+          });
+      });
+    });
+
+    describe('afterResetPasswordConfirmationPoll', () => {
+      it('notifies the channel of login, halts by default', () => {
+        sinon.spy(broker, 'send');
+
+        // customizeSync is required to send the `login` message, but
+        // it won't be set because the user hasn't visited the signup/in
+        // page.
+        account.unset('customizeSync');
+
+        return broker.afterResetPasswordConfirmationPoll(account)
+          .then((result) => {
+            assert.isTrue(broker.send.calledWith('login'));
+            assert.isTrue(broker.send.calledOnce);
+            const loginData = broker.send.args[0][1];
+            assert.isFalse(loginData.customizeSync);
+
+            assert.isTrue(result.halt);
+          });
+      });
+    });
+
+    describe('afterChangePassword', () => {
+      it('notifies the channel of change_password with the new login info', () => {
+        sinon.spy(broker, 'send');
+
+        return broker.afterChangePassword(account)
+          .then(() => {
+            assert.isTrue(broker.send.calledWith('change_password'));
+          });
+      });
+    });
+
+    describe('afterDeleteAccount', () => {
+      it('notifies the channel of delete_account', () => {
+        sinon.spy(broker, 'send');
+
+        account.set('uid', 'uid');
+
+        return broker.afterDeleteAccount(account)
+          .then(() => {
+            assert.isTrue(broker.send.calledWith('delete_account'));
+          });
       });
     });
   });
